Move inference ID update tests to AbstractSemanticMapperTestCase
